Add correction for leading edge double peak
Adding implementation for correcting double peaks
which sometimes occur in the leading edge histograms.
The algorithm finds the LS at which the double peak happens,
computes the required time offset and passes this information
to the worker to correct t and t vs TOT histograms, as well as
to the harvester to save it into the SQLite file.